Reworked Settings Activity
