+++ AndroidManifest.xml
@@ -6,7 +6,7 @@
     android:versionName="2.0" >
 
     <uses-sdk
-        android:minSdkVersion="8"
+        android:minSdkVersion="11"
         android:targetSdkVersion="18" />
 
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
@@ -73,7 +73,7 @@
             android:clearTaskOnLaunch="true"
             android:configChanges="keyboardHidden|orientation" />
         <activity
-            android:name="com.morgan.design.TeaRoundPreferences"
+            android:name="com.morgan.design.activity.preferences.SettingsActivity"
             android:clearTaskOnLaunch="true"
             android:configChanges="keyboardHidden|orientation" />
     </application>
