+++ TODO.md
@@ -4,6 +4,7 @@ Main goals
 Code changes
 - [ ] Modernise and clean code
 - [x] Set target version to latest android version (18)
+- [x] Set minimum version to android version (11)
 - [ ] Alter domain logic and clean up Brew*.java 
 - [ ] Make project work with latest version of Android
 - [ ] Update all libraries it uses, including assessing alternatives
@@ -18,6 +19,7 @@ Code changes
 - [ ] Rework groups view
 - [ ] Rework stats view
 - [ ] Rework feedback form
+- [x] Rework Settings
 - [x] Rework Google Analytics integration and what is recorded
 - [x] Remove abstract activities
