+++ res/xml/preferences.xml
@@ -1,22 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen
-	xmlns:android="http://schemas.android.com/apk/res/android">
-	<PreferenceCategory
-		android:title="General">
-		<CheckBoxPreference
-			android:title="Change Log"
-			android:defaultValue="true"
-			android:summary="Show Change Log On Version Upgrade"
-			android:key="changelog" />
-		<CheckBoxPreference
-			android:title="Timer Vibration"
-			android:defaultValue="true"
-			android:summary="Vibrate on Timer Completion"
-			android:key="timerVibration" />
-		<CheckBoxPreference
-			android:title="Google Anayltics"
-			android:defaultValue="true"
-			android:summary="Enable Google Analytics. Thanks for helping make Who's Making The Brew better"
-			android:key="googleAnalytics" />
-	</PreferenceCategory>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    <PreferenceCategory android:title="General" >
+        
+        <CheckBoxPreference
+            android:defaultValue="true"
+            android:key="changelog"
+            android:summary="Show Change Log On Version Upgrade"
+            android:title="Change Log" />
+
+        <CheckBoxPreference
+            android:defaultValue="true"
+            android:key="googleAnalytics"
+            android:summary="Enable Google Analytics. Thanks for helping make Who&apos;s Making The Brew better"
+            android:title="Google Anayltics" />
+        
+    </PreferenceCategory>
+
 </PreferenceScreen>
