+++ src/com/morgan/design/activity/preferences/SettingsActivity.java
@@ -1,0 +1,14 @@
+package com.morgan.design.activity.preferences;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class SettingsActivity extends Activity {
+
+	@Override
+	protected void onCreate(Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		getFragmentManager().beginTransaction().replace(android.R.id.content, new PreferencesFragment()).commit();
+	}
+
+}
