+++ src/com/morgan/design/utils/Prefs.java
@@ -7,18 +7,13 @@ import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
 import com.morgan.design.TeaApplication;
-import com.morgan.design.TeaRoundPreferences;
+import com.morgan.design.activity.preferences.SettingsActivity;
 
-public class PreferencesUtils {
+public class Prefs {
 
 	public static final String PREF_CHANGELOG = "changelog";
-	public static final String PREF_TIMER_VIBRATE = "timerVibration";
 	public static final String PREF_GOOGLE_ANALYTIC = "googleAnalytics";
 	public static final String PREF_APP_VERSION = "app.version";
-
-	public static final String HOUR_KEY = "SharedPreference_HourKey";
-	public static final String MINUTE_KEY = "SharedPreference_MinuteKey";
-	public static final String SECOND_KEY = "SharedPreference_SecondKey";
 
 	private static SharedPreferences getPrefs(final Context context) {
 		return PreferenceManager.getDefaultSharedPreferences(context);
@@ -30,14 +25,6 @@ public class PreferencesUtils {
 
 	public static boolean setChangelogPref(final Context context, final boolean value) {
 		return getPrefs(context).edit().putBoolean(PREF_CHANGELOG, value).commit();
-	}
-
-	public static boolean getTimerVibrationPref(final Context context) {
-		return getPrefs(context).getBoolean(PREF_TIMER_VIBRATE, true);
-	}
-
-	public static boolean setTimerVibrationPref(final Context context, final boolean value) {
-		return getPrefs(context).edit().putBoolean(PREF_TIMER_VIBRATE, value).commit();
 	}
 
 	public static boolean getGoogleAnalyticsPref(final Context context) {
@@ -57,7 +44,8 @@ public class PreferencesUtils {
 	}
 
 	public static void openPreferenecesActivity(final Activity activity) {
-		final Intent intent = new Intent(activity, TeaRoundPreferences.class);
+		// final Intent intent = new Intent(activity, TeaRoundPreferences.class);
+		final Intent intent = new Intent(activity, SettingsActivity.class);
 		activity.startActivityForResult(intent, TeaApplication.ACTIVITY_PREFERENCES);
 	}
